refactored signlas
