+++ larpmanager/utils/miscellanea.py
@@ -22,11 +22,15 @@ import os
 import random
 import shutil
 import zipfile
+from io import BytesIO
 from uuid import uuid4
 
 from django.conf import settings as conf_settings
+from django.core.files.base import ContentFile
+from django.db import models
 from django.shortcuts import render
-from PIL import Image
+from PIL import Image, ImageOps
+from PIL import Image as PILImage
 
 from larpmanager.models.association import Association
 from larpmanager.models.member import Badge
@@ -182,3 +186,78 @@ def get_warehouse_optionals(ctx, def_cols):
             active = 1
     ctx["optionals"] = optionals
     ctx["no_header_cols"] = json.dumps([el + active for el in def_cols])
+
+
+def rotate_vertical_photo(instance, sender):
+    try:
+        # noinspection PyProtectedMember, PyUnresolvedReferences
+        field = instance._meta.get_field("photo")
+        if not isinstance(field, models.ImageField):
+            return
+    except Exception:
+        return
+
+    f = getattr(instance, "photo", None)
+    if not f:
+        return
+
+    if _check_new(f, instance, sender):
+        return
+
+    fileobj = getattr(f, "file", None) or f
+    try:
+        fileobj.seek(0)
+        img = PILImage.open(fileobj)
+    except Exception:
+        return
+
+    img = ImageOps.exif_transpose(img)
+    w, h = img.size
+    if h <= w:
+        return
+
+    img = img.rotate(90, expand=True)
+
+    fmt = _get_extension(f, img)
+
+    if fmt == "JPEG" and img.mode in ("RGBA", "LA", "P"):
+        img = img.convert("RGB")
+
+    out = BytesIO()
+    save_kwargs = {"optimize": True}
+    if fmt == "JPEG":
+        save_kwargs["quality"] = 88
+    img.save(out, format=fmt, **save_kwargs)
+    out.seek(0)
+
+    basename = os.path.basename(f.name) or f.name
+    instance.photo = ContentFile(out.read(), name=basename)
+
+
+def _get_extension(f, img):
+    ext = os.path.splitext(f.name)[1].lower()
+    fmt = (img.format or "").upper()
+    if not fmt:
+        if ext in (".jpg", ".jpeg"):
+            fmt = "JPEG"
+        elif ext == ".png":
+            fmt = "PNG"
+        elif ext == ".webp":
+            fmt = "WEBP"
+        else:
+            fmt = "JPEG"
+    return fmt
+
+
+def _check_new(f, instance, sender):
+    if instance.pk:
+        try:
+            old = sender.objects.filter(pk=instance.pk).only("photo").first()
+            if old:
+                old_name = old.photo.name if old.photo else ""
+                if f.name == old_name and not getattr(f, "file", None):
+                    return True
+        except Exception:
+            pass
+
+    return False
